+++ cbits/fontmanager.c
@@ -53,6 +53,11 @@ int fmCreateFont(FMcontext* ctx, const char* name, const char* filename)
 	return fonsAddFont(ctx->fs, name, filename, 0);
 }
 
+int fmCreateFontMem(FMcontext* ctx, const char* name, unsigned char* data, int dataSize)
+{
+	return fonsAddFontMem(ctx->fs, name, data, dataSize, 1, 0);
+}
+
 void fmSetScale(FMcontext* ctx, float scale) {
 	ctx->scale = scale;
 }