+++ src/Monomer/Widgets/Composite.hs
@@ -8,7 +8,7 @@ Portability : non-portable
 
 Composite widget. Main glue between all the other widgets, also acts as the main
 app widget. Composite allows to split an application into reusable parts without
-the need to implement a lower level widget. It can comunicate with its parent
+the need to implement a lower level widget. It can communicate with its parent
 component by reporting events.
 
 Requires two functions:
@@ -626,7 +626,9 @@ compositeMerge comp state wenv newComp oldComp = newResult where
   reducedResult
     | useNewRoot = toParentResult comp newState wenv styledComp tmpResult
     | otherwise = resultNode oldComp
-  !newResult = handleWidgetIdChange oldComp reducedResult
+  !newResult = reducedResult
+    & handleUserSizeReqChange wenv oldComp
+    & handleWidgetIdChange oldComp
 
 -- | Dispose
 compositeDispose